% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad.R
\name{ad}
\alias{ad}
\title{Constructor for "ad" class}
\usage{
ad(df, ignore_col = NA)
}
\arguments{
\item{df}{The data frame to build an applicability domain from}

\item{ignore_col}{Columns that will not be considered in the
applicability domain, given as a character vector.
This will likely constitute the response variable.
The default is \code{ignore_col = NA}.
These columns will be noted in the returned \code{"ad"} object.}
}
\value{
An object of the S3 class \code{"ad"}. See \verb{[find_ad()]}
for usage of this object.
}
\description{
\code{ad} creates an object of the class \code{"ad"}, which can be used to
help in defining the applicability domain of a model or to
determine X-outliers.
}
\details{
\code{ad} should be called on a data frame of training values.
Columns to be ignored should be passed to the object as well
(this will likely include the response variable or any other
identifiers that are not relevant in computation.)
}
