% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Build a model with feature selection and tuning}
\usage{
build_model(model_method, ...)
}
\arguments{
\item{model_method}{The method used. See \verb{[eval_model()]} for options.}

\item{...}{Additional arguments to pass to model-building. Includes the
parameters to test as well as details for resampling. Can also include
controls for \verb{[caret::rfe()]} passed as objects created from
\verb{[caret::rfeControl()]} to the argument \code{rfe_ctrl}.}
}
\value{
An object of S3 class \code{"tune"}. See \verb{[tune()]} for more details.
}
\description{
\code{build_model} integrates \code{caret}'s recursive feature elimination with tuning
from \verb{[tune()]}.
}
\details{
This tends to be very time-consuming. It is only implemented for \code{"rf"}. It
is suggested to run recursive feature elimination for all model methods once
and use those predictors for future model-building.
}
