% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_scale_zero.R
\name{center_scale_zero}
\alias{center_scale_zero}
\title{Center and scale data and remove predictors with near-zero variance}
\usage{
center_scale_zero(df, ignore_col = NA, quiet = T, ...)
}
\arguments{
\item{df}{The data frame to be processed.}

\item{ignore_col}{Columns that will not be preprocessed, given as a
character vector. This will likely constitute the response variable.
The default is \code{ignore_col = NA}.
These columns will be reappended to the data frame at the beginning.}

\item{quiet}{Whether to suppress the message for how many predictors are
removed in preprocessing. The default is \code{quiet = T}.}

\item{...}{Additional arguments to be passed to \code{caret::nearZeroVar}}
}
\value{
A data frame with columns centered and scaled and with variables
with near-zero variance removed. Columns can be ignored and not be
preprocessed.
}
\description{
\code{center_scale_zero} preprocesses data by applying three common
preprocessing treatments.
The data is centered so that predictors have a mean of 0. The data is
scaled so the standard deviation is 1. The standard deviation will be
corrected using \verb{[sd_pop()]}.
}
\details{
Because the data frame to be passed to the function will likely include
columns that do not need to be transformed (like columns for identification)
or response variables, there is an option to ignore these columns using
\code{ignore_col}. The input to \code{ignore_col} should be a character vector.
}
