% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sdf.R
\name{combine_sdf}
\alias{combine_sdf}
\title{Combine SDF files}
\usage{
combine_sdf(mol_dir)
}
\arguments{
\item{mol_dir}{The directory containing the SDFs to combine.}
}
\value{
An SDF as a data frame with one column.
}
\description{
\code{combine_sdf} combines a directory of SDFs into a single SDF.
For each molecule in the single SDF, the molecule name is the filename.
}
