% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnld_sdf.R
\name{dwnld_sdf}
\alias{dwnld_sdf}
\title{Download molecule structure}
\usage{
dwnld_sdf(mol, path, file_format)
}
\arguments{
\item{mol}{Name of molecule}

\item{path}{Name of directory to save file}

\item{file_format}{Type of file to download. E.g., "SDF"}
}
\value{
Data frame corresponding to whether a file downloaded, received a warning, or received an error.
}
\description{
Queries the NCI Chemical Identifier Resolver for a structure file corresponding to the name given.
File saved in the directory provided as the path.
}
