% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{eval_model}
\alias{eval_model}
\title{Evaluate a model using k-fold cross-validation}
\usage{
eval_model(
  df,
  resp = NA,
  method,
  nfold = 10,
  simplify = T,
  ignore_col = NA,
  ...
)
}
\arguments{
\item{df}{The data frame to train the model on}

\item{resp}{The name of the column to be used as a response variable.}

\item{method}{The method to be used in model-building. See the description
for available methods.}

\item{nfold}{The number of folds to use in k-fold cross-validation. The
default is \code{nfold = 10}.}

\item{simplify}{Whether to return the results from all folds or return a data
frame summarizing the results (only reporting the average and standard
deviation of all folds). The default is \code{simplify = T}.}

\item{ignore_col}{Columns that will not be used in model-building, given as a
character vector. This may be an identifying column. The default is
\code{ignore_col = NA}.}

\item{...}{Additional arguments to pass to the model method}
}
\value{
A data frame of the results of k-fold cross-validation with the
specified model parameters. Can be unsimplified (\code{simplify = F}), returning
the result on each fold, or simplified, returning the average and standard
deviation only. The function uses \code{caret::defaultSummary} and provides the
summary statistics of R-squared, RMSE, and MAE.
}
\description{
\code{eval_model} uses k-fold cross-validation to assess the performance of a type
of model on a dataset.
}
\details{
Currently, the function can evaluate the following model types, passed
through the parameter \code{method}:
\itemize{
\item \code{"rf"}: Random Forest from the package \code{randomForest}
\item \code{"linear_svm"}" SVM with linear kernel from the package \code{e1071}
}
}
