% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model_rep.R
\name{eval_model_rep}
\alias{eval_model_rep}
\title{Evaluate a model with replication}
\usage{
eval_model_rep(nrep = 10, ...)
}
\arguments{
\item{nrep}{The number of replications to perform.}

\item{...}{Additional arguments to pass to \verb{[eval_model()]}.}
}
\value{
A list with two arguments: \code{result_summary}, containing the results as
summarized by fold and replication, and \code{result_all}, containing the results
summarized by replication only.
}
\description{
\code{eval_model_rep} is similar to \verb{[eval_model()]} but it allows for multiple
evaluations. Though it is mostly a wrapper for \verb{[eval_model()]} and
\code{replicate}, the function adds additional ability to analyze results by
returning the standard deviation of model performance across replications
in addition to standard deviation across folds.
}
\details{
To make results reproducible, set a random seed before \code{eval_model_rep} using
\code{set.seed}.
}
