% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_col_outlier.R
\name{find_col_outlier}
\alias{find_col_outlier}
\title{Find outliers in a column of a data frame}
\usage{
find_col_outlier(df, col, threshold = 3)
}
\arguments{
\item{df}{The data frame to search through}

\item{col}{The name or index of the column in the data frame.}

\item{threshold}{Indices of values that are farther from the mean
than the \code{threshold} times the  deviation are removed. The
standard deviation is obtained using \verb{[sd()]}. The default value
is \code{threshold = 3}.}
}
\value{
A vector of indices where outliers occur
}
\description{
\code{find_col_outlier} finds the outliers in a column of a data frame.
It is a wrapper for \verb{[find_outlier()]}.
}
