% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outlier.R
\name{find_outlier}
\alias{find_outlier}
\title{Find outliers in a vector}
\usage{
find_outlier(x, threshold = 3)
}
\arguments{
\item{x}{A numeric vector of values}

\item{threshold}{Indices of values that are farther from the mean
than the \code{threshold} times the  deviation are removed. The
standard deviation is obtained using \verb{[sd()]}. The default value
is \code{threshold = 3}.}
}
\value{
A numeric vector of the indices of outliers.
}
\description{
\code{find_outlier} returns the index of a vector where outliers occur.
Outliers are defined as values that exceed the number of standard
deviations passed to the function as a \code{threshold} value.
}
\details{
The function\verb{[sd()]} is used to obtain the population standard
deviation. This function will remove NA and NaN values.
}
