% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ka_to_delG.R
\name{Ka_to_delG}
\alias{Ka_to_delG}
\title{Ka to dG conversion}
\usage{
Ka_to_delG(ka)
}
\arguments{
\item{ka}{The Ka value}
}
\value{
Gibbs free energy, in kJ/mol
}
\description{
\code{Ka_to_delG} converts a value from Ka (constant of association) to
Gibbs free energy (kJ/mol).
}
