% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_extval.R
\name{make_extval}
\alias{make_extval}
\title{Make an external validation set}
\usage{
make_extval(df, p = 0.15, seed = as.integer(Sys.time()))
}
\arguments{
\item{df}{The data frame to split}

\item{p}{The proportion of data in the split. Default is \verb{0.15.}}

\item{seed}{The random seed for the data. The default is the
current time as an integer, i.e., \code{as.integer(Sys.time())}.}
}
\value{
A list where the entry \code{trn} is the retained data and
the entry \code{extval} is set aside for external validation.
}
\description{
\code{make_extval} takes a data frame and splits it into a data set
for model tuning and training as well as another data set for
external validation of the model.
}
