% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regex.R
\name{make_regex}
\alias{make_regex}
\title{Convert to regex}
\usage{
make_regex(s)
}
\arguments{
\item{s}{The string to convert to regex}
}
\value{
A string with special characters escaped
}
\description{
Escapes special characters in a string for compatability with regex
}
