% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_summary_stat.R
\name{normalize_summary_stat}
\alias{normalize_summary_stat}
\title{Find the best performing model in a list}
\usage{
normalize_summary_stat(tune_list, stat_name)
}
\arguments{
\item{tune_list}{A list of \code{"tune"} objects}

\item{stat_name_list}{A character vector of which statistics to consider. The
default is to consider all statistics, provided through the character vector
\code{c("MAE", "Rsquared", "RMSE")}.}
}
\value{
An integer index of the list where the best model occurs. Uses
\code{which.max}, so if multiple maximums occur, the first will be returned.
If there are issues with finding the maximum, the default return is \code{1}.
}
\description{
\code{best_summary_stat} returns the index of an element in a list of \code{"tune"}
objects with the best combination of performance statistics.
}
\details{
This function is a helper for \verb{[tune_helper()]}.

The function relies on \verb{[caret::defaultSummary()]}. It calls on
\code{normalize_summary_stat}, which rescales the summary statistics of the list
of objects so all statistics given to the function can be evaluated together.
}
