% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.tune.R
\name{predict.tune}
\alias{predict.tune}
\title{Make predictions using the S3 "tune" object}
\usage{
predict.tune(tune_obj, ...)
}
\arguments{
\item{tune_obj}{The object for prediction}

\item{...}{Additional arguments to be passed to \code{predict}}
}
\description{
\code{predict.tune()} is a wrapper for using predict on the object stored under
the name \code{"model"} in the object.
}
