% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tune.R
\name{print.tune}
\alias{print.tune}
\title{Print an object of S3 class "tune"}
\usage{
\method{print}{tune}(tune_obj)
}
\arguments{
\item{tune_obj}{The object to be printed}
}
\description{
Calling \code{print} on a \code{"tune"} object provides details on the model type and
the model performance.
}
