% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_desc.R
\name{read_desc}
\alias{read_desc}
\title{Read a file of chemical descriptors}
\usage{
read_desc(path)
}
\arguments{
\item{path}{The path to the file to be read}
}
\description{
\code{read_desc} can read an RDS, CSV, or Excel file with chemical
descriptors. The function is a helpful wrapper for \code{readRDS},
\code{read.csv} and \code{read_excel} from the package \code{readxl}.
}
\details{
The path passed to the function must include the file extension.
}
