% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_desc_helper.R
\name{read_desc_helper}
\alias{read_desc_helper}
\title{Read a file of chemical descriptors (helper)}
\usage{
read_desc_helper(path, quiet = F, shorten = T)
}
\arguments{
\item{path}{The location of the file to be read}

\item{quiet}{Whether the function will produce messages.
The default is \code{FALSE}.}

\item{shorten}{Whether to message the shortened file name.
The default is \code{TRUE}.}
}
\description{
\code{read_desc_helper} is a wrapper for \verb{[read_desc()]} that runs the
function with error handling. It is used as the function call
in \verb{[read_desc_list()]} to prevent the function from exiting
if a file cannot be read.
}
