% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_col_na.R
\name{remove_col_na}
\alias{remove_col_na}
\title{Remove columns based on proportion of NA values}
\usage{
remove_col_na(df, threshold = 0.2, quiet = T)
}
\arguments{
\item{df}{The data frame where columns are removed}

\item{threshold}{The maximum allowable proportion of \code{NA}.
The default is \code{threshold = 0.2}}

\item{quiet}{Whether to suppress the display the number of removed columns.
The default is \code{quiet = T}.}
}
\value{
A data frame where columns whose proportion of \code{NA}s exceed a given
proportion removed
}
\description{
\code{remove_col_na} removed columns in a data frame where the proportion of
\code{NA} values exceeds the given threshold.
}
