% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_xoutlier.R
\name{remove_xoutlier}
\alias{remove_xoutlier}
\title{Remove X-outliers from data}
\usage{
remove_xoutlier(df, quiet = T, ...)
}
\arguments{
\item{df}{The data frame where X-outliers will be removed. This
should be training data only.}

\item{quiet}{Whether to message how many observations were found
to be X-outliers. The default is \code{quiet = T}.}
}
\value{
A data frame with X-outliers removed.
}
\description{
\code{remove_xoutlier} is a wrapper for \verb{[predict.ad()]} and indexing the
data frame to retain values that are not X-outliers.
}
\details{
Additional arguments that may be useful to pass are \verb{ignore_col = }
to specify columns to ignore in building the applicability domain
object using \verb{[ad()]} inside the function.

\code{remove_xoutlier} is similar to \verb{[remove_ad()]}. The latter function
is for removing observations outside the applicability domain of the
QSAR for testing data. It requires an additional argument \code{ad} that
specifies the \code{"ad"} class object to define the applicability domain
of the training data.
}
