% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nan.R
\name{replace_nan}
\alias{replace_nan}
\title{Replace NaNs in a data frame with NA}
\usage{
replace_nan(df)
}
\arguments{
\item{df}{The data frame where values will be replaced}
}
\value{
A data frame with \code{NA} in place of \code{NaN}
}
\description{
\code{replace_nan} replaces the not-a-number values in a dataframe with \code{NA}s.
This makes operations that default to using \code{na.rm} or similar arguments
easier to troubleshoot.
}
