% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retain_name.R
\name{retain_name}
\alias{retain_name}
\title{Retain names in lapply}
\usage{
retain_name(x, f, ...)
}
\arguments{
\item{x}{The list to iterate over}

\item{f}{The function to apply}

\item{...}{Additional arguments to pass to \code{f}}
}
\value{
A list of the same length as \code{x} that is the output of
\code{lapply} with the names preserved
}
\description{
\code{retain_name} is a wrapper for \code{lapply} that retains the names
of the list passed to the appropriate function.
The function call is nearly identical to that of \code{lapply}.
}
