% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_pop.R
\name{sd_pop}
\alias{sd_pop}
\title{Obtain standard deviation}
\usage{
sd_pop(x, quiet = T)
}
\arguments{
\item{x}{Vector of numeric values}

\item{quiet}{Whether to print a message if NAs or NaNs exist}
}
\value{
A numeric value corresponding to the population standard
deviation of the input.
}
\description{
\code{sd_pop} retrieves the population standard deviation of a vector of observations.
The function will automatically remove NA values.
}
