% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train.R
\name{split_train}
\alias{split_train}
\title{Create \code{num} splits in training data saved in a single list}
\usage{
split_train(trn, nfolds, seed)
}
\arguments{
\item{seed}{The random seed, for reproducibility}

\item{num}{The number of folds created}
}
\value{
A list containing \code{num} folds of the training set data
}
\description{
Create \code{num} splits in training data saved in a single list
}
