% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_svm.R
\name{train_svm}
\alias{train_svm}
\title{Output SVM model results over multiple folds}
\usage{
train_svm(folds, kernel, param, fold_avg = T)
}
\arguments{
\item{folds}{A list of folds (training data)}

\item{kernel}{The SVM kernel}

\item{param}{A list of the parameters that can be passed to the SVM. Requires that the list names are the model's arguments.}

\item{fold_avg}{Default TRUE. Returns the average RMSE and R2 for the set of folds. Set as F to retain individual performance.}
}
\value{
Either a data frame of the averaged results or a data frame of the individual performance on each fold.
}
\description{
Output SVM model results over multiple folds
}
