% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_svm_fold.R
\name{train_svm_fold}
\alias{train_svm_fold}
\title{A helper for \code{train_svm}.}
\usage{
train_svm_fold(fold, name, parameters)
}
\arguments{
\item{fold}{The fold containing the data}

\item{name}{The fold number}

\item{parameters}{A list of the parameters that can be passed to the SVM. Requires that the list names are the model's arguments.}
}
\value{
The performance of an SVM with the listed parameters on a single fold of data.
}
\description{
A helper for \code{train_svm}.
}
