% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.randomForest.R
\name{tune.randomForest}
\alias{tune.randomForest}
\title{Tune a random forest}
\usage{
\method{tune}{randomForest}(method, df, resp, nfold = 10, nrep = 1, ignore_col = NA, ...)
}
\arguments{
\item{method}{The model-building method. Should be \code{"rf"} at this point.}

\item{df}{The data frame to train on}

\item{resp}{The name of the column containing the response variable}

\item{nfold}{The number of folds to use in evaluation. Default is \code{10}.}

\item{nrep}{The number of repetitions to use in evaluation. Default is \code{1}.}

\item{ignore_col}{Columns to ignore during model-building. Default is \code{NA}.}
}
\value{
An object of S3 class \code{"tune"}.
}
\description{
\code{tune.randomForest} tests the function \verb{[randomForest::randomForest()]} with
provided parameters.
}
