% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_helper.R
\name{tune_helper}
\alias{tune_helper}
\title{A helper function for \code{tune}}
\usage{
tune_helper(method, df, resp, nfold, nrep, ignore_col, ...)
}
\arguments{
\item{method}{The model-building method}

\item{df}{The data frame to train on}

\item{resp}{The name of the column containing the response variable}

\item{nfold}{The number of folds to use in evaluation}

\item{nrep}{The number of repetitions to use in evaluation}

\item{ignore_col}{Columns to ignore during model-building}
}
\value{
A list containing several fields relevant for the model:
\itemize{
\item \verb{$param}: The parameters used in the best model
\item \verb{$result_summary}: The summarized results for the best model
\item \verb{$result_all}: The results for each repetition for the best model
\item \verb{$best_index}: The index of the best model relative to \verb{$all_tune}
\item \verb{$all_tune}: The list of all parameters used in tuning
}
}
\description{
\code{tune_helper} is a helper function that summarizes part of the tuning process
for the method \code{tune}.
}
\details{
Calls on \code{best_summary_stat} to find the best-performing model.
}
