% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sdf.R
\name{write_sdf}
\alias{write_sdf}
\title{Write an SDF}
\usage{
write_sdf(sdf, path)
}
\arguments{
\item{sdf}{The SDF as a data frame. Accepts the output data frame
from \code{combine_sdf}}

\item{path}{The file path for the SDF. If it does not end with
".SDF", the appropriate file suffix will be applied.}
}
\value{
A data frame in the format of an SDF.
}
\description{
\code{write_sdf} can write the output of \code{combine_sdf} into an SDF file.
The function is a wrapper for \code{write.table} with settings adjusted
for valid SDF files.
The path passed to the function does not have to be suffixed with
".SDF". The function will append the correct file suffix.
}
