% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_scale.R
\name{center_scale}
\alias{center_scale}
\title{Center and scale data and remove predictors with near-zero variance}
\usage{
center_scale(df, ignore_col = NA, return_ad = F, ad_obj = NA, quiet = T, ...)
}
\arguments{
\item{df}{The data frame to be processed.}

\item{ignore_col}{Columns that will not be preprocessed, given as a
character vector. This will likely constitute the response variable.
The default is \code{ignore_col = NA}.
These columns will be reappended to the data frame at the beginning.}

\item{return_ad}{Whether or not to return the applicability domain object.
Set this to \code{return_ad = T} if this is training data and the resulting
applicability domain object could be useful for testing data.}

\item{quiet}{Whether to return a message if there are columns in the original
data frame that are dropped.}

\item{ad}{An optional \code{"ad"} object to pass. Use this if you are centering
and scaling test data. Default is \code{ad = NA} and a new object is created
and discarded within the function (unless \code{return_ad = T}).}
}
\value{
If \code{return_ad = F} (default), a data frame with columns centered
and scaled.
Columns can be ignored and not be preprocessed.
If \code{return_ad = T}, a list with the data frame with columns centered
and scaled as well as the \code{"ad"} class object.
These are labeled \code{"df"} and \code{"ad"}, respectively.
Note that only the columns in \code{ignore_col} and named in the \code{ad_obj} provided
will be returned. Set \code{quiet = F} to be notified if columns are lost.
}
\description{
\code{center_scale} preprocesses data.
}
\details{
The data is centered so that predictors have a mean of 0. The data is
scaled so the standard deviation is 1. The standard deviation will be
corrected using \verb{[sd_pop()]}.

After the above cleaning steps, the mean of each descriptor should be 0.
The NAs in the data frame will be replaced with 0, using the assumption
that missing values can be estimated to be the expectation of the descriptor.

This uses an \code{"ad"} S3 object constructed using \verb{[ad()]}.

Because the data frame to be passed to the function will likely include
columns that do not need to be transformed (like columns for identification)
or response variables, there is an option to ignore these columns using
\code{ignore_col}. The input to \code{ignore_col} should be a character vector.
}
