% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnld_mol_list.R
\name{dwnld_mol_list}
\alias{dwnld_mol_list}
\title{Download multiple molecule structures}
\usage{
dwnld_mol_list(mol_list, path, file_format = "SDF")
}
\arguments{
\item{mol_list}{Character vector of molecule names}

\item{path}{Name of directory to save file}

\item{file_format}{Type of file to download. Default is "SDF".}
}
\value{
Data frame corresponding to whether a file downloaded, received a warning, or received an error.
}
\description{
Queries the NCI Chemical Identifier Resolver for structure
files corresponding to the names given in a vector.
File saved in the directory provided as the path.
}
