% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regex.R
\name{make_regex}
\alias{make_regex}
\title{Convert to regex}
\usage{
make_regex(s)
}
\arguments{
\item{s}{The string to convert to regex}
}
\value{
A string with special characters escaped
}
\description{
\code{make_regex} escapes special regex characters in a string. The function
addresses the grouping characters (parentheses, brackets, and hyphens),
matching for the beginning and end of a string ("^" and "$"), regex
for number of repetitions, as well as the dot "." for any character.
}
\details{
The function assumes that the string passed to the function needs to
found as-is. That is, the string is not indicative of any regex pattern
matching.
}
