% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_desc_list.R
\name{read_desc_list}
\alias{read_desc_list}
\title{Read a list of chemical descriptor files}
\usage{
read_desc_list(path_list, quiet = F, shorten = T)
}
\arguments{
\item{path_list}{The list of locations of chemical descriptors}

\item{quiet}{Whether the function will produce messages.
The default is \code{FALSE}.}

\item{shorten}{Whether to message the shortened file name.
The default is \code{TRUE}.}
}
\description{
\code{read_desc_list} is a wrapper for \verb{[read_desc()]} and \code{lapply}.
The function calls \verb{[read_desc_helper()]} for additional error
handling. Acceptable files are RDS, CSV, and Excel files.
}
