% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune}
\alias{tune}
\title{Tune parameters of a model-building method}
\usage{
tune(method, ...)
}
\arguments{
\item{method}{The method to be used in model-building. See the description
for \verb{[eval_model()]} for available methods.}

\item{...}{Additional arguments to be passed to model-building. This will
likely be vectors of the values of the parameters to test.}
}
\value{
An object of the S3 class \code{"tune"}. Includes a list of the model with
the best performing parameters.
\itemize{
\item \verb{$model}: the final model with the tuned parameters
\item \verb{$param_tested}: a list of the parameters used in the tuning process
\item \verb{$nfold_tested}: the number of folds in each iteration of tuning
\item \code{nrep_tested}: the number of repetitions in each iteration of tuning
\item \code{pred_name}: the predictors from the data set
}
}
\description{
\code{tune} evaluates model performance on a combination of parameters. The methods
available are the same as \verb{[eval_model()]}.
}
\details{
Calling \code{print} on a \code{"tune"} object provides details on the model type and
the model performance.
}
